Worked on it a bit
So, for now everything works on current internet browsers. Aside from a minor issue i`m trying to resolve, we can move on to the about,etc,etc portions of the sections. also IE is not functioning, but i`ll resolve that at a later date. For now, we move on.