Prepping up for late work
The scroll is acting a bit weird so will need to look into it.
Otherwise, its now filling the page with projects/descriptions, finalizing the bottom seabed, animating the sea, animating the links.

In hindsight, i should have focused on a clear project than this, since i was not aware my schedule at work would be so hectic around the spring/summer, but i`ll going to finish this. And hopefully go back to doing backend. The front-end experience is great and should serve to aid me, but its really a commitment at times.